Release 0.9.0
Total new framework to support the upcoming Warfronts,
Broker/Minimap Icon changes to Horde / Alliance based upon who's currently in control of the selected Warfront.
Added a new broker text option to show the status from both Factions.
Added option to hide the menu when in combat.
"Hide Known Item" now has a whitelist to still show Rare's with a certain kind of drop which you already obtained, in both the "Menu Options" and the "Woldmap Icons".
Added option to show only the current or all Warfront Status in either the menu or it's own popup.
Added "Auto Change" options (Not yet available as they are function which only works with multiple Warfronts) to automatically change the menu and/or save it as your "Current Tracked Zone" to the Warfront you just entered.
Added a highly customisable sorting system for the menu. your can sort on name, type or drop and select how you want to group them together.
Added color options for the Rare's depending on their type.
Added a "Compact Mode" to the UnitFrame Integration.
Added configurable message and sound when hitting max level.
Added a broadcast message and configurable sound when one of the warfronts changes from one Faction to the other.