Release 0.9.4
- Minor tweaks.
- Fixed Orwell Stevenson's Mount ID.
- Merged both Menu's and Worldmap Icon's "Hide Options" into 1 "Master Filter", which options can be found by it's own menu in the Options Window.
Both Menu and Worldmap Icons uses this new filter by default but can be overridden in the config.
At login the new Master Filter will inherit the options set in the Hide Section of the Worldmap Icons.
Added an option to filter out Rare's with an unknown drop in both the Master Filter and the unit's own "Hide" options when overriding the Master Filter.
- Changed the Status Window to Hoover over the Zone Name instead of the Title.