Release 0.9.8
- Implemented Warfront "Timewalk" Phase detector which is set enabled by default and overrules the "Only show for level 120" option.
With this option enabled, Icons only show on the World and Minimap when in the correct Phase.
When you talk to a "Time Traveler NPC" in the area to return to the past the Icons will disappear.
Talk again to go to the present time enables the Icons again.
This feature overrules "Show Only at Level 120" as low level characters always are in the 'past' phase of the zone, so no Icons will be shown.
- Fixed an issue with the Icons not refreshing after a Warfront changes to another Faction while you're in-game.