Release 0.9.9
- Added 2 new options i n the "TomTom" section.
"Announce In Party chat" let's you automatically print the information about which Rare you're currently going to. When enabled it's broadcasting and also sets your TomTom Waypoint when you click an entry in the Menu or on the Worlmap Icons.
Note: If you need to re-broadcast the message in Party Chat you can "Shift-Left Click" them to avoid creating another Waypoint
- "Only When Group Leader": does what it says. In case the leader doesn't has this Addon installed you can uncheck this option to still send the Coordinates when clicking a Menu Entry or Worldmap Icons or just "Shift-Left Click" it to only broadcast.
- Implemented option "Show Footer in Window" in Master Filter to disable a few lines with default texts in the popup windows.
- And again some small adjustments. I'm thinking about not even naming this anymore as each update always has some altered functions, optimisations and fixes for bugs no-one may even notice.