Release 1.0.0
It took a while as I was ill for quite some time. I'm going to pickup on where I was with this Addon and implement a windows who stays on your screen for the ones who want that feature. This will take more time to implement and as I'm still on the recovery side of my illness I can't put a date on it yet but I try to work on it as much as I can.
For now I have an update for the "Party Announce" system (Shift-click an icon on the map or a rare in the menu to shoutout where you're heading to so your group gets a message).
- Added check for what type of group. If in normal party announce will be in Party Chat, if in raid group announce will be in Raid Chat.
- A few stability changes which makes my addon eat less memory then it did before, and it wasn't too bad before so now it runs even smoother.