Release 1.0.2
Added Rare's and most of it's loot from the Mechangon zone.
Added Rare's and most of it's loot from the Nazjatar zone.
Added filter to filter out "Quest Only" Rare's. This is for now mainly a use for the menu filter as I still have to implement a system to check if the quest has been completed and shows it that way. It's in the make and this is a way to disable them in the menu for now if you're willing to!

So for now you have a fully functional tracker for kills and loot, with all know customisations you can make. I still have to implement some more features but they will come soon. Expect an update or 2 every week, more if there is a nasty bug of course!

Still need to add:

System to check if Quest Drops are Completed.
More drops from Rare's when found.
System to check the Status of the Zone.