Release 1.0.3
Added scroll to the menu.
Implemented a shout-out. If you Shift Right-click a Rare in the menu or the icon on the Worldmap, it wil automatically say the rare is up and it's coordinates in local chat (/1).
Added the option to filer out "Item Only" Rare's in both Master Filter and the separate Menu/Worldmap filter. This way everything can be filtered out again.
Updated a lot of information about the Rares. Fixed some locations, added Cave Icons for the ones inside a cave and more. This closes issue #12, #13, #14 and #15. New reports can be made in an Issue I've created where you can leave comments about wrong info. It's more convenient to have it all in 1 thread.
Fixed a bug with the Loot Window for Cave Icons. They wil now display the correct information.

More features, like tracking the Quest Items will be implemented in the comming weekend/week. I have to take things slowly but I have nothing scheduled for next week so I'm sure I have time to fix that part!