Release 1.0.4
- Implemented checking if a Quest or Blueprint (also a quest item) item is completed. If completed they're marked complete and the filter handles the rest as you've configured it to react.
- Added two categories for the drops in the Settings, "Blueprints" and "Quest". Both are added to the filter so you can hide/whitelist/idk with it as you're used to in my settings.
- Checked every Rare in Mechagon to ensure my loot table is up-to-date with the latest reports on WoWhead. Nazjatar is the next one I'll organise. Whit this I've also include everything from user input at my issue page.

Next on the list is sorting out Nazjatar. I also want to get the menu to only show a drop name once when the Rare drops multiple items that has the same category. As said I can't do too much but I'm happy to get you guys this updated version. This way I can offer you the new filter settings to filter out unwanted Rare's and an up-to-date loot table for Mechagon. I'll hope to have another update ready in a couple of days.