Release 1.2.4
- Updated the database and assigned the right roles for all Shadowland Rares respectively to they HP
- Assign some Rare's and Loot to a specific Covenant as some are only Interact-able/Loot-able if you have the same Covenant.
- For the Rare's and Loot above, they are hidden by default (expect when  Rare drops at least 1 non Covenant Bound item), but can be show with the 2 new options in the Filters.
Added filter to show All Rare's, regardless of their Covenant Bound. And an option to show non Covenant Bound Rare's who only drop Covenant Bound loot. It might be possible for them to still drop an Anima Scroll so that's the reason why I implemented this. Both options can of course be set in the Mast Filter (which handles both the Menu and the Worldmap Icons) and the separated Menu and Worldmap Icons Filter.
Note: If you use any of these options some loot will be displayed in red words in the Menu. This indicates the drop is for a different Covenant than you signed up for and will not drop.
- Implemented a check for Transmog Drops. Once learned they will be marked completed.
- The loot window will now display which Covenant a Rare or Loot is is assigned for. The text is in green if you can use it or red if not. As I do offer a lot of color setting I'd might add an option to change these and some more colors but that's not on my High Priority list.
- As most Rare's drops a handful of 5 Anima Scrolls I changed all Unknown drops to Anima Only.